+++ apps/common/hooks/useBalance.ts
@@ -1,6 +1,4 @@
-import {useMemo} from 'react';
 import {toAddress} from '@yearn-finance/web-lib/utils/address';
-import {toNormalizedBN} from '@yearn-finance/web-lib/utils/format.bigNumber';
 import {useWallet} from '@common/contexts/useWallet';
 
 import type {TAddress, TDict} from '@yearn-finance/web-lib/types';
@@ -8,8 +6,7 @@ import type {TNormalizedBN} from '@yearn-finance/web-lib/utils/format.bigNumber'
 
 export function useBalance({
 	address,
-	chainID,
-	source
+	chainID
 }: {
 	address: string | TAddress;
 	chainID: number;
@@ -17,12 +14,5 @@ export function useBalance({
 }): TNormalizedBN {
 	const {getBalance} = useWallet();
 
-	const balance = useMemo((): TNormalizedBN => {
-		if (source) {
-			return source?.[toAddress(address)] || toNormalizedBN(0);
-		}
-		return getBalance({address: toAddress(address), chainID: chainID});
-	}, [source, getBalance, address, chainID]);
-
-	return balance;
+	return getBalance({address: toAddress(address), chainID: chainID});
 }