+++ apps/vaults/Wrapper.tsx
@@ -13,7 +13,10 @@ export function Wrapper({children, router}: {children: ReactElement; router: Nex
 		<div className={'mx-auto my-0 max-w-6xl pt-4 md:mb-0 md:mt-16'}>
 			<Meta meta={manifest} />
 			<AppSettingsContextApp>
-				<WalletForZapAppContextApp>{children}</WalletForZapAppContextApp>
+				<WalletForZapAppContextApp
+					chainID={router?.query?.chainID ? Number(router?.query?.chainID) : undefined}>
+					{children}
+				</WalletForZapAppContextApp>
 			</AppSettingsContextApp>
 		</div>
 	);